+++ src/main/java/com/week07/hanghaeinside/domain/post/dto/PostResponseDto.java
@@ -18,8 +18,8 @@ public class PostResponseDto {
     private String postImg;
     private LocalDateTime createAt;
     private int viewCnt;
-    private Long heartCnt;
-    private Long unHeartCnt;
+    private int heartCnt;
+    private int unHeartCnt;
